Support expressions with cell references.
