+++ src/shared/cellHelpers.js
@@ -1,3 +1,5 @@
+import { isNumber } from 'substance'
+
 export function getCellState(cell) {
   return cell.state
 }
@@ -7,6 +9,7 @@ export function isExpression(source) {
 }
 
 export function getCellValue(cell) {
+  if (!cell) return undefined
   if (cell.state) {
     return cell.state.value
   } else {
@@ -58,4 +61,44 @@ export function valueFromText(preferredType, text) {
   }
 }
 
-export { getCellLabel } from '../sheet/sheetHelpers'
+export const ALPHABET = "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
+
+export function getColumnLabel(colIdx) {
+  if (!isNumber(colIdx)) {
+    throw new Error('Illegal argument.')
+  }
+  var label = ""
+  while(true) { // eslint-disable-line
+    var mod = colIdx % ALPHABET.length
+    colIdx = Math.floor(colIdx/ALPHABET.length)
+    label = ALPHABET[mod] + label
+    if (colIdx > 0) colIdx--
+    else if (colIdx === 0) break
+  }
+  return label
+}
+
+export function getCellLabel(rowIdx, colIdx) {
+  let colLabel = getColumnLabel(colIdx)
+  let rowLabel = rowIdx + 1
+  return colLabel + rowLabel
+}
+
+export function getColumnIndex(colStr) {
+  let index = 0
+  let rank = 1
+  for (let i = 0; i < colStr.length; i++) {
+    let letter = colStr[i]
+    index += rank * ALPHABET.indexOf(letter)
+    rank++
+  }
+  return index
+}
+
+export function getRowCol(cellId) {
+  var match = /^([A-Z]+)([1-9][0-9]*)$/.exec(cellId)
+  return [
+    parseInt(match[2], 10)-1,
+    getColumnIndex(match[1])
+  ]
+}
